requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
python-dotenv>=1.0.0
Pillow>=10.2.0
gunicorn>=21.2.0
flake8>=7.0.0
django-cors-headers>=4.3.1
whitenoise>=6.6.0
